requirements.txt
asgiref==3.8.1
Django==4.2.17
django-debug-toolbar==5.0.1
django-jalali==7.4.0
django-jalali-date==2.0.0
django-phonenumber-field==8.0.0
Faker==37.0.0
jalali_core==1.0.0
jdatetime==5.2.0
persiantools==5.2.0
phonenumbers==9.0.0
pillow==11.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
python-decouple==3.8
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
